Migrate authn tests
